+++ valhalla/accounts/serializers.py
@@ -2,7 +2,6 @@ from rest_framework import serializers
 from django.contrib.auth.models import User
 
 from valhalla.accounts.models import Profile
-from valhalla.common.configdb import ConfigDB
 
 
 class ProfileSerializer(serializers.ModelSerializer):
@@ -32,15 +31,11 @@ class UserSerializer(serializers.ModelSerializer):
         ]
 
     def get_available_instrument_types(self, obj):
-        telescope_classes = set()
         instrument_types = set()
-        configdb = ConfigDB()
         for proposal in obj.proposal_set.filter(active=True):
             for timeallocation in proposal.timeallocation_set.all():
-                telescope_classes.add(timeallocation.telescope_class)
-        for telescope_class in telescope_classes:
-            for instrument_type in configdb.get_active_instrument_types({'telescope': telescope_class}):
-                instrument_types.add(instrument_type)
+                if timeallocation.instrument_name:
+                    instrument_types.add(timeallocation.instrument_name)
         return list(instrument_types)
 
     def get_tokens(self, obj):